refactor: initializeMovie 및 searchMovie 함수 분리 및 내부 로직 분리
